Create Scope object and implement watchers and digest loop
Create scope object that stores list of watchers and a reference to the last watcher that was dirty.
Add Scope. prototype that takes in a watch function, and listener function as well as a boolean to check equality by reference or value.  The watch function simply returns an item off of scope, the value of which is to be compared against the last value. The listener function is executed if the value has changed. Creating a new  adds a watcher to the front of the 55694watchers array.
Add Scope. prototype that executes all the watchers on scope, at least once, and continues executing them while at least one value in the previous iteration changed, or until the loop has iterated 10 times, on which it throws an error.
Add Scope.55694digestOnce prototype that is used by Scope..  This executes all the watchers and checks if the value has changed. If it has the watcher's listener function is called. This handles short circuiting if the last dirty watcher on the scope comes back clean.
Add Scope.55694areEqual prototype that takes in 2 values and determines if they are equal (by reference unless the boolean passed in is true, then they are checked by value). NaNs are handled correctly